% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catTest.R
\name{catTestfisher}
\alias{catTestfisher}
\title{Fisher's exact test for use with \code{\link{summary_factorlist}}}
\usage{
catTestfisher(tab)
}
\arguments{
\item{tab}{A frequency table such as that created by \code{table()}.}
}
\description{
Fisher's exact test for use with \code{\link{summary_factorlist}}
}
\examples{
explanatory = c("sex.factor", "obstruct.factor")
dependent = "mort_5yr"
colon_s \%>\%
  summary_factorlist(dependent, explanatory, p = TRUE, catTest = catTestfisher)
}
