% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_package.R
\docType{data}
\name{colon_s}
\alias{colon_s}
\title{Chemotherapy for Stage B/C colon cancer}
\format{
A data frame with 929 rows and 33 variables
}
\source{
\code{\link[survival]{colon}}
}
\usage{
data(colon_s)
}
\description{
This is a modified version of \code{survival::\link[survival]{colon}}.These
are data from one of the first successful trials of adjuvant chemotherapy for
colon cancer. Levamisole is a low-toxicity compound previously used to treat
worm infestations in animals; 5-FU is a moderately toxic (as these things go)
chemotherapy agent. There are two records per person, one for recurrence and
one for death
}
\keyword{data}
