% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphmixed.R
\name{coxphmixed}
\alias{coxphmixed}
\title{Cox proprotional hazards mixed effects models: \code{finalfit} model wrapper}
\usage{
coxphmixed(.data, dependent, explanatory, random_effect)
}
\arguments{
\item{.data}{Data frame.}

\item{dependent}{Character vector of length 1:  name of survival object in
form \code{Surv(time, status)}.}

\item{explanatory}{Character vector of any length: name(s) of explanatory
variables.}

\item{random_effect}{Character vector of length 1, either, (1) name of random
intercept variable, e.g. "var1", (automatically convered to "(1 | var1)");
or, (2) the full \code{lme4} specification, e.g. "(var1 | var2)". Note
parenthesis MUST be included in (2) but NOT included in (1).}
}
\value{
A multivariable \code{\link[coxme]{coxme}} fitted model. Output is of
  class \code{coxme}.
}
\description{
Using \code{finalfit} conventions, produces multivariable Cox Proportional
Hazard regression mixed effects models for a set of explanatory variables and
random effects structure against a survival object.
}
\details{
Uses \code{\link[coxme]{coxme}} with \code{finalfit} modelling conventions.
Output can be passed to \code{\link{fit2df}}. This function is in development
given the coxme package is new/
}
\examples{
# Cox Proportional Hazards mixed effects analysis.
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "Surv(time, status)"
random_effect = "hospital"
colon_s \%>\%
	coxphmixed(dependent, explanatory, random_effect)
	
}
\seealso{
\code{\link{fit2df}, \link{finalfit_merge}}

Other finalfit model wrappers: 
\code{\link{coxphmulti}()},
\code{\link{coxphuni}()},
\code{\link{crrmulti}()},
\code{\link{crruni}()},
\code{\link{glmmixed}()},
\code{\link{glmmulti_boot}()},
\code{\link{glmmulti}()},
\code{\link{glmuni}()},
\code{\link{lmmixed}()},
\code{\link{lmmulti}()},
\code{\link{lmuni}()},
\code{\link{svyglmmulti}()},
\code{\link{svyglmuni}()}
}
\concept{finalfit model wrappers}
