% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphuni.R
\name{coxphuni}
\alias{coxphuni}
\title{Cox proprotional hazards univariable models: \code{finalfit} model wrapper}
\usage{
coxphuni(.data, dependent, explanatory)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  name of survival object in form \code{Surv(time, status)}.}

\item{explanatory}{Character vector of any length: name(s) of explanatory variables.}
}
\value{
A list of univariable \code{\link[survival]{coxph}} fitted model outputs.
  Output is of class \code{coxphlist}.
}
\description{
Using \code{finalfit} conventions, produces multiple univariable Cox Proportional Hazard
  regression models for a set of explanatory variables against a survival object.
}
\details{
Uses \code{\link[survival]{coxph}} with \code{finalfit} modelling conventions. Output can be
  passed to \code{\link{fit2df}}.
}
\examples{
# Cox Proportional Hazards univariable analysis.
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "Surv(time, status)"
colon_s \%>\%
	coxphuni(dependent, explanatory) \%>\%
	fit2df()
}
\seealso{
\code{\link{fit2df}}

Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}
}
