% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrmulti.R
\name{crrmulti}
\alias{crrmulti}
\title{Competing risks multivariable regression: \code{finalfit} model wrapper}
\usage{
crrmulti(.data, dependent, explanatory, ...)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{dependent}{Character vector of length 1:  name of \code{ftime} and
\code{fstatus} in form \code{c("ftime", "fstatus")}.}

\item{explanatory}{Character vector of any length: name(s) of explanatory
variables.}

\item{...}{Other arguments to \code{\link[cmprsk]{crr}}}
}
\value{
A multivariable \code{\link[cmprsk]{crr}} fitted model
  class \code{crr}.
}
\description{
Using \code{finalfit} conventions, produces multivariable Competing Risks
Regression models for a set of explanatory variables.
}
\details{
Uses \code{\link[cmprsk]{crr}} with \code{finalfit} modelling conventions.
Output can be passed to \code{\link{fit2df}}.
}
\examples{
library(dplyr)
melanoma = boot::melanoma
melanoma = melanoma \%>\% 
  mutate(
    status_coxph = ifelse(status == 2, 0, # "still alive"
      ifelse(status == 1, 1, # "died of melanoma"
        0)), # "died of other causes is censored" 
    status_crr = ifelse(status == 2, 0, # "still alive"
      ifelse(status == 1, 1, # "died of melanoma"
        2)), # "died of other causes" 
    sex = factor(sex),
    ulcer = factor(ulcer)
  )

dependent_coxph = c("Surv(time, status_coxph)")			
dependent_crr = c("Surv(time, status_crr)")
explanatory = c("sex", "age", "ulcer")
melanoma \%>\% 
  summary_factorlist(dependent_crr, explanatory, column = TRUE, fit_id = TRUE) \%>\% 
  ff_merge(
    melanoma \%>\% 
      coxphmulti(dependent_coxph, explanatory) \%>\% 
      fit2df(estimate_suffix = " (Cox PH multivariable)")
    ) \%>\% 
  ff_merge(
    melanoma \%>\% 
      crrmulti(dependent_crr, explanatory) \%>\% 
      fit2df(estimate_suffix = " (competing risks)")
    ) \%>\% 
  select(-fit_id, -index) \%>\% 
  dependent_label(melanoma, dependent_crr)
}
\seealso{
\code{\link{fit2df}, \link{finalfit_merge}}

Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{crruni}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}
}
\concept{\code{finalfit} model wrappers}
