% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{dependent_label}
\alias{dependent_label}
\title{Make a label for the dependent variable}
\usage{
dependent_label(.data, dependent)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  quoted name of depdendent
variable. Can be continuous, a binary factor, or a survival object of form
\code{Surv(time, status)}}
}
\value{
Returns "Depdendent: " and the name of the depdendent.
}
\description{
Not called directly.
}
\keyword{internal}
