% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{error_colon_fct_levels}
\alias{error_colon_fct_levels}
\title{Errors: colon in factor levels}
\usage{
error_colon_fct_levels(.data)
}
\arguments{
\item{.data}{Data frame.}
}
\value{
Logical
}
\description{
Errors: colon in factor levels
}
\keyword{internal}
