% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract variable labels and names}
\usage{
extract_labels(.data)
}
\arguments{
\item{.data}{Data frame.}
}
\value{
A data frame with three columns: first (vname), variabe names; second
  (vlabel), variables labels; third (vfill), variable labels and when null
  variable names.
}
\description{
Extract variable labels and names
}
\examples{
colon_s \%>\%
  extract_labels()
}
\keyword{internal}
