% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{extract_variable_label}
\alias{extract_variable_label}
\title{Extract variable labels from dataframe}
\usage{
extract_variable_label(.data)
}
\arguments{
\item{.data}{Dataframe containing labelled variables.}
}
\description{
Variable labels can be created using \code{\link{ff_label}}. Some functions
strip variable labels (variable attributes), e.g. \code{forcats::fct_recode}.
Use this function to create a vector of variable labels from a data frame.
Then use \code{\link{ff_relabel}} to relabel variables in data frame.
}
\examples{
colon_s \%>\%
  extract_variable_label
}
