% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{ff_describe}
\alias{ff_describe}
\title{psych:: describe interval only}
\usage{
ff_describe(.data, na.rm = TRUE, interp = FALSE, skew = TRUE,
  ranges = TRUE, trim = 0.1, type = 3, check = TRUE, fast = NULL,
  quant = NULL, IQR = FALSE, omit = FALSE)
}
\arguments{
\item{.data}{Dataframe or matrix.}

\item{na.rm}{Logical. Remove missing data.}

\item{interp}{Logical. Interpolate median.}

\item{skew}{Logical. Include skew and kurtosis.}

\item{ranges}{Logical. Include range.}

\item{trim}{Drop the top and bottom trim fraction}

\item{type}{Skew or kurtosis}

\item{check}{Logical. Include non-numeric variables}

\item{fast}{Not used.}

\item{quant}{Specify quantiles.}

\item{IQR}{Logical. Interquartile range.}

\item{omit}{Logical. Omit non-numerical variables.}
}
\value{
Dataframe.
}
\description{
Internal, not normally called directly, though can be.
}
\references{
https://github.com/cran/psych/blob/master/R/describe.R
}
\author{
Procedures for Psychological, Psychometric, and Personality Research.
  Homepage: https://personality-project.org/r/psych
  https://personality-project.org/r/psych-manual.pdf
}
\keyword{internal}
