% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{ff_eval}
\alias{ff_eval}
\title{Eval for `lm` and `glm` model wrappers}
\usage{
ff_eval(.)
}
\arguments{
\item{.}{}
}
\description{
Internal function, not called directly. This is in reponse to a long running
issue of the best way to pass `weights` to `lm()` and `glm()`. See here
https://stackoverflow.com/questions/54383414/passing-weights-to-glm-using-rlang
}
\keyword{internal}
