% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{ff_label}
\alias{ff_label}
\alias{finalfit_label}
\title{Label a variable}
\usage{
ff_label(.var, variable_label)

finalfit_label(.var, variable_label)
}
\arguments{
\item{.var}{Quoted variable name}

\item{variable_label}{Quoted variable label}
}
\value{
Labelled variable
}
\description{
Label a variable
}
\examples{
colon_s$sex.factor \%>\%
  ff_label("Sex") \%>\%
  str()
}
\seealso{
\code{\link{extract_variable_label}} \code{\link{ff_relabel}}
}
