% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{ff_parse_formula}
\alias{ff_parse_formula}
\title{Parse a formula to finalfit grammar}
\usage{
ff_parse_formula(.formula)
}
\arguments{
\item{.formula}{an object of class "formula" (or one that can be coerced to that class).}
}
\value{
A list containing dependent, explanatory and random effects variables
}
\description{
Parse a formula to finalfit grammar
}
\examples{
ff_parse_formula(mort ~ age + sex + (1 | hospital))
}
