% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_remove_ref.R
\name{ff_percent_only}
\alias{ff_percent_only}
\alias{finalfit_percent_only}
\title{Include only percentages for factors in \code{\link{summary_factorlist}} output}
\usage{
ff_percent_only(.data)

finalfit_percent_only(.data)
}
\arguments{
\item{.data}{Output from \code{\link{finalfit}} or similar.}
}
\value{
Data frame.
}
\description{
Include only percentages for factors in \code{\link{summary_factorlist}} output
}
\examples{
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = 'mort_5yr'
colon_s \%>\%
  summary_factorlist(dependent, explanatory) \%>\%
  ff_percent_only()
}
