% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{ff_relabel_df}
\alias{ff_relabel_df}
\alias{finalfit_relabel_df}
\title{Relabel variables from data frame after tidyverse functions}
\usage{
ff_relabel_df(.data, .df)

finalfit_relabel_df(.data, .df)
}
\arguments{
\item{.data}{Data frame or tibble after applicaton of label stripping functions.}

\item{.df}{Original data frame which contains labels.}
}
\value{
Data frame or tibble
}
\description{
Relabel variables from data frame after tidyverse functions
}
