% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_remove_ref.R
\name{ff_remove_p}
\alias{ff_remove_p}
\alias{finalfit_remove_p}
\title{Remove p-value from output}
\usage{
ff_remove_p(.data)

finalfit_remove_p(.data)
}
\arguments{
\item{.data}{Output from \code{\link{finalfit}} or similar.}
}
\value{
Data frame.
}
\description{
This will work with \code{\link{finalfit}} and any \code{\link{fit2df}}
output.
}
\examples{
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = 'mort_5yr'
colon_s \%>\%
  finalfit(dependent, explanatory) \%>\%
  ff_remove_p()
}
