% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{ff_stratify_helper}
\alias{ff_stratify_helper}
\title{Help making stratified summary_factorlist tables}
\usage{
ff_stratify_helper(df.out, .data)
}
\arguments{
\item{df.out}{Output from \code{summary_factorlist}}

\item{.data}{Original data frame used for \code{summary_factorlist}.}
}
\description{
Help making stratified summary_factorlist tables
}
\examples{
library(dplyr)
explanatory = c("age.factor", "sex.factor")
dependent = "perfor.factor"

# Pick option below
split = "rx.factor"
split = c("rx.factor", "node4.factor")

# Piped function to generate stratified crosstabs table
colon_s \%>\%
  group_by(!!! syms(split)) \%>\% #Looks awkward, but avoids unquoted var names
  group_modify(~ summary_factorlist(.x, dependent, explanatory)) \%>\%
  ff_stratify_helper(colon_s)
}
