% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_package.R
\docType{package}
\name{finalfit-package}
\alias{finalfit-package}
\title{finalfit: Quickly create elegant final results tables and charts when modelling.}
\description{
finalfit: Quickly create elegant final results tables and charts when modelling.
}
\section{finalfit functions}{

The foo functions ...
}

