% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_coxph.R
\name{fit2df.coxph}
\alias{fit2df.coxph}
\title{Extract \code{survival::coxph} model fit results to dataframe: \code{finalfit} model extracters}
\usage{
\method{fit2df}{coxph}(fit, condense = TRUE, metrics = NULL,
  estimate_suffix = "", ...)
}
\arguments{
\item{fit}{Output from \code{survival::coxph} models.}

\item{condense}{Logical: when true, effect estimates, confidence intervals and p-values
are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted (not yet implemented).}

\item{estimate_suffix}{Character vector of length one specifying string to be appended
to output column name}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters. When \code{metrics=TRUE} output is a list of two dataframes,
  one is model parameters, one is model metrics. length two
}
\description{
Takes output from \code{survival::\link[survival]{coxph}} and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df.coxph} is the model extract method for \code{survival::\link[survival]{coxph}}.
}
\examples{
library(finalfit)
library(dplyr)
library(survival)

fit = coxph(Surv(time, status) ~ age.factor + sex.factor + obstruct.factor + perfor.factor,
  data = colon_s)

fit \%>\%
	fit2df(estimate_suffix=" (multivariable)")
}
\seealso{
Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}

Other \code{finalfit} model extractors: \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}}
}
