% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_glm.R
\name{fit2df.glm}
\alias{fit2df.glm}
\title{Extract \code{glm} model fit results to dataframe: \code{finalfit} model
extracters}
\usage{
\method{fit2df}{glm}(.data, condense = TRUE, metrics = FALSE,
  remove_intercept = TRUE, explanatory_name = "explanatory",
  estimate_name = "OR", estimate_suffix = "", p_name = "p",
  digits = c(2, 2, 3), confint_sep = "-", ...)
}
\arguments{
\item{.data}{Output from \code{finalfit} model wrappers.}

\item{condense}{Logical: when true, effect estimates, confidence intervals and p-values
are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted.}

\item{remove_intercept}{Logical: remove the results for the intercept term.}

\item{explanatory_name}{Name for this column in output}

\item{estimate_name}{Name for this column in output}

\item{estimate_suffix}{Appeneded to estimate name}

\item{p_name}{Name given to p-value estimate}

\item{digits}{Number of digits to round to (1) estimate, (2) confidence
interval limits, (3) p-value.}

\item{confint_sep}{String to separate confidence intervals, typically "-" or
" to ".}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters. When \code{metrics=TRUE} output is a
  list of two dataframes, one is model parameters, one is model metrics.
}
\description{
Takes output from \code{\link[stats]{glm}} and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df.glm} is the model extract method for standard
\code{\link[stats]{glm}} models, which have not used \code{finalfit} model
wrappers.
}
\examples{
library(finalfit)
library(dplyr)

fit = glm(mort_5yr ~  age.factor + sex.factor + obstruct.factor + perfor.factor,
  data=colon_s, family="binomial")

fit \%>\%
  fit2df(estimate_suffix=" (multivariable)")
}
\seealso{
Other \code{finalfit} model extracters: \code{\link{fit2df.glmboot}}
}
