% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_glmboot.R
\name{fit2df.glmboot}
\alias{fit2df.glmboot}
\title{Extract \code{glmboot} model fit results to dataframe: \code{finalfit} model extracters}
\usage{
\method{fit2df}{glmboot}(fit, condense = TRUE, metrics = NULL,
  estimate_suffix = "", ...)
}
\arguments{
\item{fit}{Output from \code{\link{glmmulti_boot}}.}

\item{condense}{Logical: when true, effect estimates, confidence intervals and p-values
are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted (not currently implemented).}

\item{estimate_suffix}{Character vector of length one specifying string to be appended
to output column name}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters.
}
\description{
Takes output from \code{\link[stats]{glm}} with \code{boot::\link[boot]{boot}}
confidence intervals on fixed effect coefficients and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df.glmboot} is the model extract method for \code{\link{glmmulti_boot}} models.
}
\examples{
library(finalfit)
library(dplyr)
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "mort_5yr"

colon_s \%>\%
	glmmulti_boot(dependent, explanatory) \%>\%
	fit2df(estimate_suffix=" (multivariable (BS CIs))")

}
\seealso{
Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}

Other \code{finalfit} model extracters: \code{\link{fit2df.glm}}
}
