% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_lmerMod.R
\name{fit2df.lmerMod}
\alias{fit2df.lmerMod}
\title{Extract \code{lmerMod} model fit results to dataframe: \code{finalfit} model
extracters}
\usage{
\method{fit2df}{lmerMod}(fit, condense = TRUE, metrics = FALSE,
  estimate_suffix = "", ...)
}
\arguments{
\item{fit}{Output from \code{lme4::\link[lme4]{lmer}} or
\code{finalfit::\link{lmmixed}}.}

\item{condense}{Logical: when true, effect estimates, confidence intervals
and p-values are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted (nor
currently implemented).}

\item{estimate_suffix}{Character vector of length one specifying string to be
appended to output column name}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters.
}
\description{
Takes output from \code{lme4::\link[lme4]{lmer}} and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df.lmerMod} is the model extract method for standard
\code{lme4::\link[lme4]{lmer}} models and for the
\code{finalfit::\link{lmmixed}} model wrapper.
}
\examples{
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
random_effect = "hospital"
dependent = "nodes"

colon_s \%>\%
  lmmixed(dependent, explanatory, random_effect) \%>\%
	 fit2df(estimate_suffix=" (multilevel")
}
\seealso{
Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}

Other \code{finalfit} model extractors: \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}}
}
