% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_lmlist.R
\name{fit2df.lmlist}
\alias{fit2df.lmlist}
\title{Extract \code{lmuni} and \code{lmmulti} model fit results to dataframe:
\code{finalfit} model extracters}
\usage{
\method{fit2df}{lmlist}(fit, condense = TRUE, metrics = FALSE,
  estimate_suffix = "", ...)
}
\arguments{
\item{fit}{Output from \code{finalfit} model wrappers.}

\item{condense}{Logical: when true, effect estimates, confidence intervals
and p-values are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted.}

\item{estimate_suffix}{Character vector of length one specifying string to be
appended to output column name}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters. When \code{metrics=TRUE} output is a
  list of two dataframes, one is model parameters, one is model metrics.
  length two
}
\description{
Takes output from \code{finalfit} model wrappers and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df.lmlist} is the model extract method for \code{lmuni} and
\code{lmmulti}.
}
\examples{
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "nodes"

colon_s \%>\%
  lmmulti(dependent, explanatory) \%>\%
  fit2df()

}
\seealso{
Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}

Other \code{finalfit} model extractors: \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}}
}
