% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmuni.R
\name{glmuni}
\alias{glmuni}
\title{Binomial logistic regression univariable models: \code{finalfit} model wrapper}
\usage{
glmuni(.data, dependent, explanatory)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  name of depdendent variable (must have 2 levels).}

\item{explanatory}{Character vector of any length: name(s) of explanatory variables.}
}
\value{
A list of univariable \code{\link[stats]{glm}} fitted model outputs.
  Output is of class \code{glmlist}.
}
\description{
Using \code{finalfit} conventions, produces multiple univariable binomial logistic
  regression models for a set of explanatory variables against a binary dependent.
}
\details{
Uses \code{\link[stats]{glm}} with \code{finalfit} modelling conventions. Output can be
  passed to \code{\link{fit2df}}.
}
\examples{
library(finalfit)
library(dplyr)
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "mort_5yr"

colon_s \%>\%
	glmuni(dependent, explanatory) \%>\%
	fit2df(estimate_suffix=" (univariable)")


}
\seealso{
\code{\link{fit2df}}

Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{lmmixed}},
  \code{\link{lmmulti}}, \code{\link{lmuni}}
}
