% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_plot.R
\name{hr_plot}
\alias{hr_plot}
\title{Produce a hazard ratio table and plot}
\usage{
hr_plot(.data, dependent, explanatory, dependent_label = NULL,
  factorlist = NULL, coxfit = NULL, breaks = NULL,
  column_space = c(-0.5, 0, 0.5), ...)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  name of survival object in form \code{Surv(time, status)}.}

\item{explanatory}{Character vector of any length: name(s) of explanatory variables.}

\item{dependent_label}{Label on plot for survival object, e.g. "Overall survival".}

\item{factorlist}{Option to provide output directly from \code{\link{summary_factorlist}()}.}

\item{coxfit}{Option to provide output directly from \code{coxphmulti()}.}

\item{breaks}{Manually specify x-axis breaks in format \code{c(0.1, 1, 10)}.}

\item{column_space}{Adjust table column spacing.}

\item{...}{Other parameters.}
}
\value{
Returns a table and plot produced in \code{ggplot2}.
}
\description{
Produce hazard ratio table and plot from a Cox Proportional Hazards analysis, \code{survival::coxph()}.
}
\examples{
# HR plot
library(finalfit)
library(dplyr)

data(colon_s)
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "Surv(time, status)"
colon_s \%>\%
  hr_plot(dependent, explanatory, dependent_label = "Survival")

}
\seealso{
\code{\link{fit2df}}

Other finalfit wrappers: \code{\link{finalfit_merge}},
  \code{\link{or_plot}}, \code{\link{summary_factorlist}},
  \code{\link{surv_plot}}
}
