% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{interp.median}
\alias{interp.median}
\title{Interpolate median from psych}
\usage{
interp.median(x, w = 1, na.rm = TRUE)
}
\description{
Interpolate median from psych
}
\author{
Procedures for Psychological, Psychometric, and Personality Research.
  Homepage: https://personality-project.org/r/psych
  https://personality-project.org/r/psych-manual.pdf
}
\keyword{internal}
