% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{interp.quantiles}
\alias{interp.quantiles}
\title{Interpolate quantiles from psych}
\usage{
interp.quantiles(x, q = 0.5, w = 1, na.rm = TRUE)
}
\description{
Interpolate quantiles from psych
}
\author{
Procedures for Psychological, Psychometric, and Personality Research.
  Homepage: https://personality-project.org/r/psych
  https://personality-project.org/r/psych-manual.pdf
}
\keyword{internal}
