% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{is.survival}
\alias{is.survival}
\title{Test character describes survival object}
\usage{
is.survival(.name)
}
\arguments{
\item{.name}{Character string to test}
}
\value{
Logical
}
\description{
Test character describes survival object
}
\examples{
var_s = "Surv(mort, time)"
is.survival(var_s) #TRUE
var_s = "Sur(mort, time)"
is.survival(var_s) #FALSE
}
\keyword{internal}
