% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{kurtosi}
\alias{kurtosi}
\title{kurtosi from psych}
\usage{
kurtosi(x, na.rm = TRUE, type = 3)
}
\description{
kurtosi from psych
}
\author{
Procedures for Psychological, Psychometric, and Personality Research.
  Homepage: https://personality-project.org/r/psych
  https://personality-project.org/r/psych-manual.pdf
}
\keyword{internal}
