% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{labels_to_column}
\alias{labels_to_column}
\title{Labels to column names}
\usage{
labels_to_column(.data)
}
\arguments{
\item{.data}{Data frame or tibble.}
}
\value{
Data frame or tibble
}
\description{
Labels to column names
}
\examples{
library(dplyr)
colon_s \%>\% 
  select(sex.factor) \%>\% 
  labels_to_column()
}
