% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{labels_to_level}
\alias{labels_to_level}
\title{Labels to level}
\usage{
labels_to_level(.data, .labels)
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{.labels}{Output from \code{extract_variable_label}.}
}
\value{
Data frame or tibble
}
\description{
For use with forcats::fct_relabel.
}
\examples{
library(dplyr)
vlabels = extract_variable_label(colon_s)
colon_s \%>\%
 select(sex.factor, obstruct.factor) \%>\% 
 tidyr::gather() \%>\% 
 mutate(
  key = forcats::fct_relabel(key, labels_to_level, vlabels)
 )
}
