% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_plot.R
\name{missing_df}
\alias{missing_df}
\title{Missing values data frame}
\usage{
missing_df(.data, dependent = NULL, explanatory = NULL)
}
\arguments{
\item{.data}{Data frame.}

\item{dependent}{Character vector. Optional name of dependent variable.}

\item{explanatory}{Character vector. Optional name(s) of explanatory
variables.}
}
\value{
Data frame of missing values for all variables.
}
\description{
Create a data frame of missing vs. observed values for all variables
provided. \code{Dependent} and \code{explanatory} are for convenience and are
optional.
}
\examples{
colon_s \%>\%
  missing_df()
}
