% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_glimpse.R
\name{missing_glimpse}
\alias{missing_glimpse}
\title{Summary of missing values}
\usage{
missing_glimpse(.data, dependent = NULL, explanatory = NULL,
  digits = 1)
}
\arguments{
\item{.data}{Data frame.}

\item{dependent}{Optional character vector: name(s) of depdendent
variable(s).}

\item{explanatory}{Optional character vector: name(s) of explanatory
variable(s).}

\item{digits}{Number of decmial places to show for percentage missing.}
}
\value{
Data frame.
}
\description{
Summary of missing values
}
\examples{
colon_s \%>\%
	missing_glimpse()
}
