% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_plot.R
\name{missing_plot}
\alias{missing_plot}
\title{Missing values heatmap}
\usage{
missing_plot(.data, dependent = NULL, explanatory = NULL, title = NULL,
  plot_opts = NULL)
}
\arguments{
\item{.data}{Data frame.}

\item{dependent}{Character vector. Optional name of dependent variable.}

\item{explanatory}{Character vector. Optional name(s) of explanatory variables.}

\item{title}{Character vector. Optional title for plot}

\item{plot_opts}{A list of arguments to be appended to the ggplot call by "+".}
}
\value{
Heat map of missing values in dataset.
}
\description{
Missing values heatmap
}
\examples{

colon_s \%>\%
  missing_plot()
}
