% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{p_tidy}
\alias{p_tidy}
\title{Round p-values but keep trailing zeros}
\usage{
p_tidy(x, digits, prefix = "=")
}
\arguments{
\item{x}{Numeric vector of values to round}

\item{digits}{Integer of length one: value to round to.}

\item{prefix}{Appended in front of values for use with \code{condense_fit}.}
}
\value{
Vector of strings.
}
\description{
Internal function, not called directly
}
\details{
e.g. for 3 decimal places I want 0.100, not 0.1. Note this function with
convert 0.000 to <0.001. All other values are prefixed with "=" by default
}
