% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{plot_title}
\alias{plot_title}
\title{Label plot title}
\usage{
plot_title(.data, dependent, dependent_label, prefix = "", suffix = "")
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  quoted name of depdendent
variable. Can be continuous, a binary factor, or a survival object of form
\code{Surv(time, status)}}

\item{prefix}{Prefix for dependent label}

\item{suffix}{Suffix for dependent label}
}
\description{
Not called directly.
}
\keyword{internal}
