% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{print.data.frame.ff}
\alias{print.data.frame.ff}
\title{Print methods for finalfit data frames}
\usage{
\method{print}{data.frame.ff}(x, ...)
}
\arguments{
\item{.data}{Data frame}
}
\value{
Data frame with no line numbers
}
\description{
Print methods for finalfit data frames
}
\keyword{internal}
