% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{remove_intercept}
\alias{remove_intercept}
\title{Remove intercept from model output}
\usage{
remove_intercept(.data, intercept_name = "(Intercept)")
}
\arguments{
\item{.data}{Numeric vector of values to round}

\item{intercept_name}{Name given to interept in model. Should never have to
change from default.}
}
\value{
Vector of strings.
}
\description{
Internal function, not called directly
}
\keyword{internal}
