% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_label.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove variable labels.}
\usage{
remove_labels(.data)
}
\arguments{
\item{.data}{Data frame}
}
\value{
The original data frame with variable label attributes removed.
}
\description{
Remove variable labels.
}
\examples{
colon_s \%>\%
  remove_labels()
}
\keyword{internal}
