% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{rm_duplicate_labels}
\alias{rm_duplicate_labels}
\title{Remove duplicate levels within \code{\link{summary_factorlist}}: \code{finalfit} helper function}
\usage{
rm_duplicate_labels(factorlist, na_to_missing = TRUE)
}
\arguments{
\item{factorlist}{A factorlist intermediary.}

\item{na_to_missing}{Logical: convert \code{NA} to 'Missing' when \code{na_include=TRUE}.}
}
\value{
Returns a \code{factorlist} dataframe.
}
\description{
Not called directly.
}
\keyword{internal}
