% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{rm_duplicates}
\alias{rm_duplicates}
\title{Remove duplicates and replace}
\usage{
rm_duplicates(.var, fromLast = FALSE, replacement = "")
}
\arguments{
\item{.var}{Vector.}

\item{fromLast}{Logical. Consider duplication from last to first.}

\item{replacement}{Character for what to replace duplicate with.}
}
\value{
Character vector.
}
\description{
Remove duplicates and replace
}
