% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{skew}
\alias{skew}
\title{Internal only}
\usage{
skew(x, na.rm = TRUE, type = 3)
}
\description{
Internal only
}
\author{
Procedures for Psychological, Psychometric, and Personality Research.
  Homepage: https://personality-project.org/r/psych
  https://personality-project.org/r/psych-manual.pdf
}
\keyword{internal}
