% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfactorlist.R
\name{summarise_categorical}
\alias{summarise_categorical}
\title{Helper function to generate the summary for a categorical variable}
\usage{
summarise_categorical(x, column, total_col)
}
\description{
Internal function, not called directly.
}
\keyword{internal}
