% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfactorlist.R
\name{summarise_continuous}
\alias{summarise_continuous}
\title{Helper function to generate the summary for a continuous variable}
\usage{
summarise_continuous(x, cont, cont_range, total_col, digits)
}
\description{
Internal function, not called directly.
}
\keyword{internal}
