% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfactorlist.R
\name{summary_factorlist3}
\alias{summary_factorlist3}
\title{Summarise a set of factors by a dependent variable}
\usage{
summary_factorlist3(.data, dependent, explanatory, cont = "mean",
  p = FALSE, na_include = FALSE, column = FALSE, total_col = FALSE,
  orderbytotal = FALSE, fit_id = FALSE, na_to_missing = TRUE,
  add_dependent_label = FALSE, dependent_label_prefix = "Dependent: ",
  dependent_label_suffix = "")
}
\description{
Internal function, not called directly.
}
\keyword{internal}
