% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{summary_mipo}
\alias{summary_mipo}
\title{Call to mice:::summary.mipo}
\usage{
summary_mipo(object, type = c("tests", "all"), conf.int = FALSE,
  conf.level = 0.95, exponentiate = FALSE, ...)
}
\description{
Not called directly.
}
\keyword{internal}
