% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{variable_type}
\alias{variable_type}
\title{Determine type/class of a variable}
\usage{
variable_type(.var)
}
\arguments{
\item{.var}{A vector, data frame column, or equivalent.}
}
\value{
One of "factor", "character", "numeric", "logical", "date".
}
\description{
Determine type/class of a variable
}
\examples{
var_d = as.Date("12.03.18", "\%d.\%m.\%y")
var_f = factor(c("yes", "no"))
var_c = c("yes", "no")
var_n = 1:10
var_l = as.logical(c("true", "false"))
variable_type(var_d)
variable_type(var_f)
variable_type(var_c)
variable_type(var_n)
variable_type(var_l)
}
\keyword{internal}
