% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_package.R
\docType{data}
\name{wcgs}
\alias{wcgs}
\title{Western Collaborative Group Study}
\format{
A data frame with 3154 observations on the following 13 variables.
  \itemize{ \item{\code{id}}{ Subject ID } \item{\code{age}}{ Age: age in
  years } \item{\code{height}}{ Height: height in inches }
  \item{\code{weight}}{ Weight: weight in pounds } \item{\code{sbp}}{
  Systolic blood pressure: mmHg } \item{\code{dbp}}{ Diastolic blood
  pressure: mmHg } \item{\code{chol}}{ Cholesterol: mg/100 ml }
  \item{\code{personality}}{Personality type/Behavior pattern: a factor with
  levels \code{A1}, \code{A2}, \code{B3}, \code{B4} }
  \item{\code{personality_2L}}{ Dichotomous personality type / behavior
  pattern: \code{A} = aggressive; \code{B} = passive }  \item{\code{ncigs0}}{
  Smoking: Cigarettes/day } \item{\code{smoking}}{ Smoking: \code{No},
  \code{Yes} } \item{\code{arcus0}}{ Corneal arcus: \code{No}, \code{Yes} }
  \item{\code{chd}}{ Coronary heart disease event: \code{No} \code{Yes}}
  \item{\code{typechd}}{ coronary heart disease is a factor with levels
  \code{No}, \code{MI_SD} (MI or sudden death), \code{Silent_MI},
  \code{Angina} } \item{\code{timechd}}{ Observation (follow up) time: Days }
  }
}
\source{
Statistics for Epidemiology by N. Jewell (2004)
}
\usage{
data(wcgs)
}
\description{
3154 healthy young men aged 39-59 from the San Francisco area were assessed
for their personality type. All were free from coronary heart disease at the
start of the research. Eight and a half years later change in this situation
was recorded.
}
\details{
The WCGS began in 1960 with 3,524 male volunteers who were employed by 11
California companies. Subjects were 39 to 59 years old and free of heart
disease as determined by electrocardiogram. After the initial screening, the
study population dropped to 3,154 and the number of companies to 10 because
of various exclusions. The cohort comprised both blue- and white-collar
employees. At baseline the following information was collected:
socio-demographic including age, education, marital status, income,
occupation; physical and physiological including height, weight, blood
pressure, electrocardiogram, and corneal arcus; biochemical including
cholesterol and lipoprotein fractions; medical and family history and use of
medications; behavioral data including Type A interview, smoking, exercise,
and alcohol use. Later surveys added data on anthropometry, triglycerides,
Jenkins Activity Survey, and caffeine use. Average follow-up continued for
8.5 years with repeat examinations
}
\references{
Coronary Heart Disease in the Western Collaborative Group Study
  Final Follow-up Experience of 8 1/2 Years Ray H. Rosenman, MD; Richard J.
  Brand, PhD; C. David Jenkins, PhD; Meyer Friedman, MD; Reuben Straus, MD;
  Moses Wurm, MD JAMA. 1975;233(8):872-877.
  doi:10.1001/jama.1975.03260080034016.
}
\keyword{data}
