% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{catTestfisher}
\alias{catTestfisher}
\title{Deprecated catTest from Hmisc for reverse dependencies}
\usage{
catTestfisher(.)
}
\arguments{
\item{.}{Null}
}
\description{
Deprecated catTest from Hmisc for reverse dependencies
}
\keyword{internal}
