% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{extract_fit}
\alias{extract_fit}
\alias{extract_fit.glm}
\alias{extract_fit.glmerMod}
\alias{extract_fit.lm}
\alias{extract_fit.lmerMod}
\alias{extract_fit.coxph}
\title{Extract model output to dataframe}
\usage{
extract_fit(...)

\method{extract_fit}{glm}(.data, explanatory_name = "explanatory",
  estimate_name = "OR", estimate_suffix = "", p_name = "p",
  confint_type = "profile", confint_level = 0.95, ...)

\method{extract_fit}{glmerMod}(.data, explanatory_name = "explanatory",
  estimate_name = "OR", estimate_suffix = "", p_name = "p",
  confint_type = "Wald", confint_level = 0.95, ...)

\method{extract_fit}{lm}(.data, explanatory_name = "explanatory",
  estimate_name = "Coefficient", estimate_suffix = "", p_name = "p",
  confint_level = 0.95, ...)

\method{extract_fit}{lmerMod}(.data, explanatory_name = "explanatory",
  estimate_name = "OR", estimate_suffix = "", p_name = "p",
  confint_type = "Wald", confint_level = 0.95, ...)

\method{extract_fit}{coxph}(.data, explanatory_name = "explanatory",
  estimate_name = "HR", estimate_suffix = "", p_name = "p", ...)
}
\arguments{
\item{...}{Other arguments.}

\item{.data}{Model output.}

\item{explanatory_name}{Name for this column in output.}

\item{estimate_name}{Name for this column in output.}

\item{estimate_suffix}{Appeneded to estimate name.}

\item{p_name}{Name given to p-value estimate}

\item{confint_type}{One of \code{c("profile", "default")} for GLM
models or \code{c("profile", "Wald", "boot")} for \code{glmer/lmer} models.
Not implemented for \code{lm, coxph or coxphlist}.}

\item{confint_level}{The confidence level required.}
}
\description{
Internal function, not usually called directly.

Internal function, not called directly.
}
\keyword{internal}
