% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract model output to dataframe}
\usage{
extract_fit(.data, explanatory_name, estimate_name, estimate_suffix, p_name,
  digits, ...)
}
\arguments{
\item{.data}{Model output.}

\item{explanatory_name}{Name for this column in output.}

\item{estimate_name}{Name for this column in output.}

\item{estimate_suffix}{Appeneded to estimate name.}

\item{p_name}{Name given to p-value estimate}

\item{digits}{Number of digits to round to (1) estimate, (2) confidence
interval limits, (3) p-value.}

\item{...}{Other arguments.}
}
\description{
Internal function, not called directly.
}
\keyword{internal}
