% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{extract_variable_label}
\alias{extract_variable_label}
\title{Extract variable labels from dataframe}
\usage{
extract_variable_label(.data)
}
\arguments{
\item{.data}{Dataframe containing labelled variables.}
}
\description{
Internal function, not called directly.
}
\keyword{internal}
