% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_package.R
\docType{package}
\name{finalfit-package}
\alias{finalfit-package}
\title{finalfit: Quickly create elegant final results tables and plots when modelling.}
\description{
Quickly create elegant final results tables and plots when modelling.
}
\section{\code{finalfit} model wrappers}{


\link{glmuni},
\link{glmmulti},
\link{glmmulti_boot},
\link{glmmixed},
\link{lmuni},
\link{lmmulti},
\link{lmmixed},
\link{coxphuni},
\link{coxphmulti},
\link{crruni},
\link{crrmulti},
\link{svyglmuni},
\link{svyglmmulti}.
}

\section{\code{finalfit} model extractor}{


Generic:
\link{fit2df}

Methods (not called directly):
\link{fit2df.glm},
\link{fit2df.glmlist},
\link{fit2df.glmboot},
\link{fit2df.lm},
\link{fit2df.lmlist},
\link{fit2df.glmerMod},
\link{fit2df.lmerMod},
\link{fit2df.coxph},
\link{fit2df.coxphlist},
\link{fit2df.crr},
\link{fit2df.crrlist},
\link{fit2df.stanfit}.
}

\section{\code{finalfit} all-in-one function}{


Generic:
\link{finalfit}.
\link{finalfit_permute}.

Methods (not called directly):
\link{finalfit.glm},
\link{finalfit.lm},
\link{finalfit.coxph}.
}

\section{\code{finalfit} plotting functions}{


\link{coefficient_plot},
\link{or_plot},
\link{hr_plot},
\link{surv_plot},
\link{ff_plot}.
}

\section{\code{finalfit} helper functions}{


\link{ff_glimpse},
\link{ff_label},
\link{ff_merge},
\link{ff_interaction}.
}

\section{\code{finalfit} prediction functions}{


\link{boot_predict},
\link{finalfit_newdata}.

Methods (not called directly):
\link{boot_compare}.
}

\section{\code{finalfit} missing data functions}{


\link{missing_glimpse},
\link{missing_pattern},
\link{missing_compare},
\link{missing_plot},
\link{missing_pairs}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewenharrison/finalfit}
  \item Report bugs at \url{https://github.com/ewenharrison/finalfit/issues}
}

}
\author{
\strong{Maintainer}: Ewen Harrison \email{ewen.harrison@ed.ac.uk}

Authors:
\itemize{
  \item Tom Drake
  \item Riinu Pius
}

}
\keyword{internal}
