% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_package.R
\docType{package}
\name{finalfit-package}
\alias{finalfit-package}
\title{finalfit: Quickly create elegant final results tables and plots when modelling.}
\description{
finalfit: Quickly create elegant final results tables and plots when modelling.
}
\section{\code{finalfit} model wrappers}{


\link{glmuni},
\link{glmmulti},
\link{glmmulti_boot},
\link{glmmixed},
\link{lmuni},
\link{lmmulti},
\link{lmmixed},
\link{coxphuni},
\link{coxphmulti}.
}

\section{\code{finalfit} model extractor}{


Generic:
\link{fit2df}

Methods (not called directly):
\link{fit2df.glm},
\link{fit2df.glmlist},
\link{fit2df.glmboot},
\link{fit2df.lm},
\link{fit2df.lmlist},
\link{fit2df.glmerMod},
\link{fit2df.lmerMod},
\link{fit2df.coxph},
\link{fit2df.coxphlist},
\link{fit2df.stanfit}.
}

\section{\code{finalfit} all-in-one function}{


Generic:
\link{finalfit}.

Methods (not called directly):
\link{finalfit.glm},
\link{finalfit.lm},
\link{finalfit.coxph}.
}

\section{\code{finalfit} plotting functions}{


\link{or_plot},
\link{hr_plot},
\link{surv_plot}.
}

\section{\code{finalfit} helper functions}{


\link{finalfit_merge},
\link{finalfit_missing}.
}

\section{\code{finalfit} prediction functions}{


\link{boot_predict},
\link{finalfit_newdata}.

Methods (not called directly):
\link{boot_compare}.
}

