% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df.R
\name{fit2df}
\alias{fit2df}
\title{Extract model fit results to dataframe (generic): \code{finalfit} model extractors}
\usage{
fit2df(.data, condense, metrics, remove_intercept, explanatory_name,
  estimate_name, estimate_suffix, p_name, digits, confint_sep, ...)
}
\arguments{
\item{.data}{Output from \code{finalfit} model wrappers.}

\item{condense}{Logical: when true, effect estimates, confidence intervals and p-values
are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted.}

\item{remove_intercept}{Logical: remove the results for the intercept term.}

\item{explanatory_name}{Name for this column in output}

\item{estimate_name}{Name for this column in output}

\item{estimate_suffix}{Appeneded to estimate name}

\item{p_name}{Name given to p-value estimate}

\item{digits}{Number of digits to round to (1) estimate, (2) confidence
interval limits, (3) p-value.}

\item{confint_sep}{String to separate confidence intervals, typically "-" or
" to ".}

\item{...}{Other arguments (not used).}
}
\value{
A dataframe of model parameters. When \code{metrics=TRUE} output is a list of two dataframes,
  one is model parameters, one is model metrics. length two
}
\description{
Takes output from \code{finalfit} model wrappers and extracts to a dataframe,
convenient for further processing in preparation for final results table.
}
\details{
\code{fit2df} is a generic (S3) function for model extract.
}
\examples{
library(finalfit)
library(dplyr)
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "mort_5yr"

colon_s \%>\%
	glmmulti(dependent, explanatory) \%>\%
	fit2df(estimate_suffix=" (univariable)")
}
\seealso{
Other \code{finalfit} model extractors: \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}}, \code{\link{fit2df.lm}},
  \code{\link{fit2df.stanfit}}
}
