% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2df_stanfit.R
\name{fit2df.stanfit}
\alias{fit2df.stanfit}
\title{Extract \code{stanfit} model fit results to dataframe: \code{finalfit} model
extracters}
\usage{
\method{fit2df}{stanfit}(fit, condense = TRUE, metrics = FALSE,
  estimate_suffix = "", X, ...)
}
\arguments{
\item{fit}{Output from \code{rstan}.}

\item{condense}{Logical: when true, effect estimates, confidence intervals
and p-values are pasted conveniently together in single cell.}

\item{metrics}{Logical: when true, useful model metrics are extracted.}

\item{estimate_suffix}{Character vector of length one specifying string to be
appended to output column name}

\item{X}{Design matrix used in Stan model.}

\item{...}{Not used.}
}
\value{
A dataframe of model parameters. When \code{metrics=TRUE} output is a
  list of two dataframes, one is model parameters, one is model metrics.
}
\description{
Takes output from \code{rstan} model and extracts to a dataframe, convenient
for further processing in preparation for final results table.
}
\details{
\code{fit2df.stanfit} is the model extract method for our standard Bayesian
hierarchical binomial logistic regression models. These models will be fully
documented separately. However this should work for a single or multilevel
Bayesian logistic regression done in Stan, as long as the fixed effects are
specified in the parameters block as a vector named \code{beta}, of length
\code{P}, where \code{P} is the number of fixed effect parameters. e.g.
parameters{
  vector[P] beta;
  }
}
\seealso{
Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmmulti}},
  \code{\link{lmuni}}

Other \code{finalfit} model extractors: \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}}, \code{\link{fit2df}}
}
