% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_internal_functions.R
\name{format_n_percent}
\alias{format_n_percent}
\title{Format n and percent as a character}
\usage{
format_n_percent(n, percent, digits)
}
\arguments{
\item{n}{Value}

\item{percent}{Value}

\item{digits}{Value}
}
\description{
Internal, function, not called directly
}
