% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmulti.R
\name{lmmulti}
\alias{lmmulti}
\title{Linear regression multivariable models: \code{finalfit} model wrapper}
\usage{
lmmulti(.data, dependent, explanatory)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector usually of length 1, but can take more than 1
dependent:  name of depdendent variable (must a continuous vector).}

\item{explanatory}{Character vector of any length: name(s) of explanatory variables.}
}
\value{
A list of multivariable \code{\link[stats]{lm}} fitted model outputs.
  Output is of class \code{lmlist}.
}
\description{
Using \code{finalfit} conventions, produces multiple multivariable linear
  regression models for a set of explanatory variables against a continuous dependent.
}
\details{
Uses \code{\link[stats]{lm}} with \code{finalfit} modelling conventions. Output can be
  passed to \code{\link{fit2df}}. Note that this function can take multiple \code{dependent}
  variables as well, but performs multiple individual models, not a multivariate analysis.
}
\examples{
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "nodes"

colon_s \%>\%
  lmmulti(dependent, explanatory) \%>\%
  fit2df()

}
\seealso{
\code{\link{fit2df}}

Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{fit2df.coxphlist}},
  \code{\link{fit2df.coxph}}, \code{\link{fit2df.glmboot}},
  \code{\link{fit2df.glmerMod}},
  \code{\link{fit2df.glmlist}}, \code{\link{fit2df.glm}},
  \code{\link{fit2df.lmerMod}},
  \code{\link{fit2df.lmlist}},
  \code{\link{fit2df.stanfit}}, \code{\link{fit2df}},
  \code{\link{glmmixed}}, \code{\link{glmmulti_boot}},
  \code{\link{glmmulti}}, \code{\link{glmuni}},
  \code{\link{lmmixed}}, \code{\link{lmuni}}
}
