% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmuni.R
\name{lmuni}
\alias{lmuni}
\title{Linear regression univariable models: \code{finalfit} model wrapper}
\usage{
lmuni(.data, dependent, explanatory)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1, name of depdendent variable (must be continuous vector).}

\item{explanatory}{Character vector of any length: name(s) of explanatory variables.}
}
\value{
A list of multivariable \code{\link[stats]{lm}} fitted model outputs.
  Output is of class \code{lmlist}.
}
\description{
Using \code{finalfit} conventions, produces multiple univariable linear
  regression models for a set of explanatory variables against a continuous dependent.
}
\details{
Uses \code{\link[stats]{lm}} with \code{finalfit} modelling conventions. Output can be
  passed to \code{\link{fit2df}}.
}
\examples{
library(finalfit)
library(dplyr)

explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "nodes"

colon_s \%>\%
  lmuni(dependent, explanatory) \%>\%
  fit2df()

}
\seealso{
\code{\link{fit2df}}

Other \code{finalfit} model wrappers: \code{\link{coxphmulti}},
  \code{\link{coxphuni}}, \code{\link{glmmixed}},
  \code{\link{glmmulti_boot}}, \code{\link{glmmulti}},
  \code{\link{glmuni}}, \code{\link{lmmixed}},
  \code{\link{lmmulti}}
}
