% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_plot.R
\name{or_plot}
\alias{or_plot}
\title{Produce an odds ratio table and plot}
\usage{
or_plot(.data, dependent, explanatory, factorlist = NULL, glmfit = NULL,
  breaks = NULL, column_space = c(-0.5, 0, 0.5), ...)
}
\arguments{
\item{.data}{Dataframe.}

\item{dependent}{Character vector of length 1:  name of depdendent variable
(must have 2 levels).}

\item{explanatory}{Character vector of any length: name(s) of explanatory
variables.}

\item{factorlist}{Option to provide output directly from
\code{\link{summary_factorlist}()}.}

\item{glmfit}{Option to provide output directly from \code{\link{glmmulti}()}
and \code{\link{glmmixed}()}.}

\item{breaks}{Manually specify x-axis breaks in format \code{c(0.1, 1, 10)}.}

\item{column_space}{Adjust table column spacing.}

\item{...}{Other parameters.}
}
\value{
Returns a table and plot produced in \code{ggplot2}.
}
\description{
Produce an odds ratio table and plot from a \code{glm()} or
\code{lme4::glmer()} model.
}
\examples{
library(finalfit)
library(dplyr
)
# OR plot
data(colon_s)
explanatory = c("age.factor", "sex.factor", "obstruct.factor", "perfor.factor")
dependent = "mort_5yr"
colon_s \%>\%
	or_plot(dependent, explanatory)

}
\seealso{
\code{\link{fit2df}}

Other finalfit wrappers: \code{\link{finalfit_merge}},
  \code{\link{hr_plot}}, \code{\link{summary_factorlist}},
  \code{\link{surv_plot}}
}
